Move {Generic,Old,New}PacketFormat out of decoder
Although the classes have decoding routines, they aren't decoders.

Note: This may be a smell that decoder routines shouldn't be in the class
themselves but in a set of external decoder classes?
