+++ encryptit/packets/__init__.py
@@ -1,1 +1,3 @@
 from .packet_type import PacketType
+from .old_format_packet_header import OldFormatPacketHeader
+from .new_format_packet_header import NewFormatPacketHeader
